# Instruction combination prototype