@maximecb 2017-07-15
Hi @krypt-n. The way you wrote combineInstrs to work as an accumulator is pretty clever. I do see some problems with it though.

The first is that you're allocating new instruction objects, which is going to have overhead. The second is that AFAIK, your design allows people to now use inc_i32, dec_i32, etc in their bytecode directly. I would prefer the combined instructions be internal only, because that gives us more freedom to change our design later. Anything we expose, people may begin to rely on.

I think I would much prefer if the loop in `compile()` checked for specific patterns, `op == "foo" && nextOp == "bar"`. At the moment, we are only looking at a few specific patterns. Helper functions can be used if it makes things simpler, as in `op == foo && nextOp(instrs, 1, "bar")`, where nextOp could check if the i+kth op is what we want.

Nitpick: curly braces go on their own line.

> the results are not overwhelming since the plush code generator wraps all operations in function calls.

I wrote it that way because every primitive op has a number of dynamic type checks. The primitive ops in Plush are actually all quite big. If we generated conditional checks for every addition, subtraction, field access, we would double or triple the size of generated code. Generating code would also take that much longer.

My goal is to eventually have inlining be done by the interpreter/JIT. By placing counters at every call site, we can detect which call sites get executed a lot, and inline specifically those function calls. We'll end up inlining just 5 or 10% in total, so there won't be a massive code size or compilation time increase, and we will remove the function call overhead where it counts.

Inlining will likely wait until after @Maaarcocr completes our first garbage collector, because both of those features are very bug-prone, and I think we should test and debug them separately.