@maximecb 2017-07-27
Hi @krypt-n. I realize that being asked to make changes to your PR isn't fun, but do you think you will complete this one? I realize that you may be busy with other things.