@krypt-n 2017-07-27
Oh, sorry. I got "distracted" with uni stuff and work. I'll implement the requested changes tomorrow.