@krypt-n 2017-07-28
I implemented the requested changes and updated this branch. Performance is not very different to the numbers in the PR description.